requirements.txt
Flask==2.3.3
requests==2.31.0
openai==1.12.0
python-dotenv==1.0.0
waitress==2.1.2
geopy==2.4.1
